pyproject.toml
[tool.poetry]
name = "product_service"
version = "0.1.0"
description = ""
authors = ["artorias <kerelkobarabash@gmail.com>"]
readme = "README.md"
packages = [
    { include = "product_service", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.12"
strawberry-graphql = {extras = ["fastapi"], version = "^0.262.6"}
pydantic-settings = "^2.8.1"
sqlalchemy = "^2.0.40"
alembic = "^1.15.2"
dishka = "^1.5.1"
asyncpg = "^0.30.0"
uvicorn = "^0.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
requests = "^2.32.3"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
